pyproject.toml
[project]
name = "interactive-gui-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.9.0",
    "mss>=10.0.0",
    "pillow>=11.2.1",
]
